feat(screenpipe-status): move screenpipe-status dialog to its own component
this way the same dialog can be triggered by user action or context if health/permissions are broken
